% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actor_time_info.R
\name{actor_pds_from_netlet}
\alias{actor_pds_from_netlet}
\title{Derive actor periods attribute from netlet}
\usage{
actor_pds_from_netlet(netlet)
}
\arguments{
\item{netlet}{multilayer netlet object}
}
\value{
data.frame formatted in the same way as the
actor_pds attribute
}
\description{
Derive actor periods attribute from netlet
}
\author{
Shahryar Minhas
}
