% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actor_time_info.R
\name{actor_pds_to_frame}
\alias{actor_pds_to_frame}
\title{actor_pds_to_frame}
\usage{
actor_pds_to_frame(netlet_actor_pds)
}
\arguments{
\item{netlet_actor_pds}{actor pds attributes from netify object}
}
\value{
a data.frame object with actor pd info in a actor-time format
}
\description{
This function converts the actor_pds attribute of netify
objects into a data.frame at the unit of observation level
}
\author{
Colin Henry, Shahryar Minhas
}
