% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_network.R
\name{add_dyad_to_network}
\alias{add_dyad_to_network}
\title{add_dyad_to_network}
\usage{
add_dyad_to_network(netlet, dyad_data_list, network_object, time = NULL)
}
\arguments{
\item{netlet}{netify object}

\item{dyad_data_list}{dyad data from netlet object}

\item{network_object}{network object to modify}

\item{time}{time indicator for longit case}
}
\value{
network object with dyad attributes added
}
\description{
Add dyad attributes to a network object from netify object
}
\author{
Shahryar Minhas
}
