% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nodal.R
\name{add_nodal}
\alias{add_nodal}
\title{Adds nodal data to a netify object}
\usage{
add_nodal(
  netlet,
  node_data,
  actor = NULL,
  time = NULL,
  node_vars = NULL,
  replace_existing = FALSE
)
}
\arguments{
\item{netlet}{a netify object}

\item{node_data}{a data.frame object}

\item{actor}{a character object indicating which variable in node_data uniquely IDs each node}

\item{time}{a character object indicating which variable in node_data tracks time}

\item{node_vars}{a vector of which variables from node_data should be merged}

\item{replace_existing}{a logical indicating whether to replace existing nodal data}
}
\value{
a netify object
}
\description{
This function takes in a data.frame and outputs a netify object.
}
\examples{

data(icews)

# cross-sectional case
icews_10 <- icews[icews$year==2010,]


verbCoop_net <- netify(
  dyad_data=icews_10,
  actor1 = 'i', actor2 = 'j', 
  symmetric=FALSE, weight='verbCoop' )

# nodal data should be a dataframe with one row per actor
# in the cross-sectional case and one row per actor per 
# time period in the longitudinal case, e.g.:
nvars = c(
  'i_polity2',
  'i_gdp', 'i_log_gdp',
  'i_pop', 'i_log_pop' )
nodeData <- unique(icews_10[,c('i', nvars)])
head(nodeData)

verbCoop_net <- add_nodal( 
  netlet=verbCoop_net, node_data=nodeData, 
  actor='i', node_vars=nvars )

# nodal data is stored in the nodal_data attribute
# as a data.frame, it can be accessed in the following way:
node_data <- attr(verbCoop_net, 'nodal_data')
head(node_data)

# longitudinal case
verbCoop_longit_net <- netify(
    dyad_data=icews, 
    actor1='i', actor2='j', time='year',
    symmetric=FALSE,
    weight='verbCoop' )

nodeData <- unique(icews[,c('i', 'year', nvars)])

verbCoop_longit_net <- add_nodal(
  netlet=verbCoop_longit_net,
  node_data=nodeData, 
  actor='i', time='year',
  node_vars = nvars )

# and in the longitudinal case, it can be accessed
# in the same way
node_data <- attr(verbCoop_longit_net, 'nodal_data')
head(node_data)

}
\author{
Colin Henry, Shahryar Minhas
}
