% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_igraph.R
\name{add_nodal_to_igraph}
\alias{add_nodal_to_igraph}
\title{add_nodal_to_igraph}
\usage{
add_nodal_to_igraph(netlet, node_data, igraph_object, time = NULL)
}
\arguments{
\item{netlet}{netify object}

\item{node_data}{node data from netlet object}

\item{igraph_object}{igraph object to modify}

\item{time}{time indicator for longit case}
}
\value{
igraph object with nodal attributes added
}
\description{
Add nodal attributes to an igraph object from netify object
}
\author{
Shahryar Minhas
}
