% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{add_var_time_check}
\alias{add_var_time_check}
\title{add_var_time_check}
\usage{
add_var_time_check(netlet, time)
}
\arguments{
\item{netlet}{user inputted netlet object}

\item{time}{user inputted time variable}
}
\value{
returns a NULL object but stops the process
if there is an error detected
}
\description{
Stops the process if the user tries to add time to a
non-longitudinal dataset and vice versa
}
\author{
Cassy Dorff, Shahryar Minhas
}
