% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_igraph.R
\name{adj_igraph_positions}
\alias{adj_igraph_positions}
\title{adj_igraph_positions}
\usage{
adj_igraph_positions(adj_mat, igraph_object)
}
\arguments{
\item{adj_mat}{adjacency matrix}

\item{igraph_object}{igraph object to modify}
}
\value{
matrix object of how actor positions match
between the adj_mat and igraph_object
}
\description{
Match igraph edge order with matrix data for the
purpose of setting edge attributes
}
\author{
Shahryar Minhas
}
