% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_plot_args.R
\name{adjust_plot_args}
\alias{adjust_plot_args}
\title{Adjust Plotting Parameters and Data for Netify Visualization}
\usage{
adjust_plot_args(plot_args, net_dfs, obj_attrs)
}
\arguments{
\item{plot_args}{A list of user-defined plotting arguments that may include options for text overlap,
edge curvature, edge angles, line segment parameters, and labeling specific actors.}

\item{net_dfs}{A list containing nodal and edge data frames as returned by \code{decompose_netlet}.
\code{nodal_data} should contain node information and \code{edge_data} should contain edge information.}

\item{obj_attrs}{Attributes of the \code{netify} object, which should include details about whether the
network is symmetric, its type (cross-sectional or longitudinal), and other relevant properties.}
}
\value{
A list containing the adjusted \code{plot_args} and modified \code{net_dfs} ready for plotting.
}
\description{
This function prepares the plotting parameters and data from a \code{netify} object for visualization.
It sets defaults for plotting parameters if not supplied by the user and modifies the node and
edge data frames based on the provided parameters and the attributes of the \code{netify} object.
}
\details{
The function checks and sets default values for various plot parameters such as:
\itemize{
\item \code{check_overlap}: Whether to check for overlap in text labels (default TRUE).
\item \code{curve_edges}: Whether to curve the edges in the plot (default FALSE).
\item \code{edge_curvature}: The curvature of the edges (default 0.5).
\item \code{edge_angle}: The angle of line segments (default 90 degrees).
\item \code{edge_ncp}: Number of control points for edge curvature (default 5).
\item \code{edge_arrow}: Default arrow type for directed edges.
It also adjusts the edge and node data based on whether the network is directed or undirected,
and handles specific labeling of actors and transparency settings based on edge weights.
}
}
\author{
Cassy Dorff, Shahryar Minhas
}
