% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{array_to_list}
\alias{array_to_list}
\title{array_to_list}
\usage{
array_to_list(arr, preserveAttr = TRUE)
}
\arguments{
\item{arr}{three dimensional array to list}

\item{preserveAttr}{logical indicating whether to preserve attributes}
}
\value{
list object
}
\description{
This function converts a three dimensional array
into a list of matrices
}
\author{
Shahryar Minhas
}
