% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_helpers.R
\name{check_layer_compatible}
\alias{check_layer_compatible}
\title{Cycle through elements of a netlet object and make sure
that they are identical before we try to merge into
a multilayer netify object}
\usage{
check_layer_compatible(a_list, elems, msg)
}
\arguments{
\item{a_list}{list object to check}

\item{elems}{character vector with names of elements to check}

\item{msg}{character vector of length two warning user
if elements are found to be not identical. First part of
vector is the preamble before the element name and
second part is what should come after the element name.}
}
\description{
Cycle through elements of a netlet object and make sure
that they are identical before we try to merge into
a multilayer netify object
}
