% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_duplicate_dyads}
\alias{count_duplicate_dyads}
\title{Determine number of duplicate dyad-time obs}
\usage{
count_duplicate_dyads(actor1, actor2, time)
}
\arguments{
\item{actor1}{character vector for actor1}

\item{actor2}{character vector for actor2}

\item{time}{numeric vector for time}
}
\value{
an integer count of number of duplicate dyads
}
\description{
Determine number of duplicate dyad-time obs
}
\author{
Shahryar Minhas
}
