% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_duplicate_dyads_indexed}
\alias{count_duplicate_dyads_indexed}
\title{Alternative implementation for small actor sets
Uses integer encoding for actors when possible}
\usage{
count_duplicate_dyads_indexed(actor1, actor2, time)
}
\arguments{
\item{actor1}{character vector for actor1}

\item{actor2}{character vector for actor2}

\item{time}{numeric vector for time}
}
\value{
an integer count of number of duplicate dyads
}
\description{
Alternative implementation for small actor sets
Uses integer encoding for actors when possible
}
\author{
Shahryar Minhas
}
