% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{df_check}
\alias{df_check}
\title{df_check}
\usage{
df_check(df, msg = "Error: check data type. `dyad_data` is not a dataframe.")
}
\arguments{
\item{df}{user inputted object to check}

\item{msg}{msg to user if df check fails}
}
\value{
data.frame object
}
\description{
Checks to make sure a data.frame is inputted and
if a tibble data.frame is inputted then it is
converted to a base R data.frame object
}
\author{
Ha Eun Choi, Shahryar Minhas
}
