% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{edge_value_check}
\alias{edge_value_check}
\title{edge_value_check}
\usage{
edge_value_check(weight, sum_dyads, time = FALSE)
}
\arguments{
\item{weight}{user inputted weight value}

\item{sum_dyads}{user inputted sum_dyads logical}

\item{time}{logical indicating whether inputted data is longitudinal}
}
\value{
returns a NULL object but provides warnings to users
}
\description{
Warns user about how edge values in adjacency matrices will be determined
}
\author{
Cassy Dorff, Shahryar Minhas
}
