% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_layers.R
\name{filter_layers}
\alias{filter_layers}
\title{Filters to specified layer(s) of a netify multilayer network object.}
\usage{
filter_layers(netlet, layer_labels)
}
\arguments{
\item{netlet}{multilayer netlet object}

\item{layer_labels}{character: label of the layers to be filtered to}
}
\value{
a netlet object subsetted to the chosen layers
}
\description{
Filters to specified layer(s) of a netify multilayer network object.
}
\examples{

library(netify)
data(icews)
icews_10 <- icews[icews$year==2010,]

# generate netify objects that will be layered together
icews_verbCoop <- netify(
    dyad_data=icews_10, actor1='i', actor2='j',
    symmetric=FALSE, weight='verbCoop' )            

icews_verbConf <- netify(
    dyad_data=icews_10, actor1='i', actor2='j',
    symmetric=FALSE, weight='verbConf' )

icews_matlCoop <- netify(
    dyad_data=icews_10, actor1='i', actor2='j',
    symmetric=FALSE, weight='matlCoop' )

icews_matlConf <- netify(
    dyad_data=icews_10, actor1='i', actor2='j',
    symmetric=FALSE, weight='matlConf' )

icews_layers <- layer_netify(
    netlet_list=list(
        icews_verbCoop, icews_matlCoop,
        icews_verbConf, icews_matlConf ),
    layer_labels=c(
        'verbCoop', 'matlCoop',
        'verbConf', 'matlConf' ) )

# filter to verbCoop and verbConf layers
icews_verb <- filter_layers(
    netlet=icews_layers, 
    layer_labels=c('verbCoop', 'verbConf') )

# so you know what to filter, 
# information on layer labels can be accessed 
# from  the `layer_labels` attribute 
attr(icews_layers, 'layers')

}
\author{
Shahryar Minhas
}
