% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gen_symm_id}
\alias{gen_symm_id}
\title{gen_symm_id}
\usage{
gen_symm_id(dyad_data, actor1, actor2, time = NULL)
}
\arguments{
\item{dyad_data}{dyadic data.frame object}

\item{actor1}{name of actor1 column}

\item{actor2}{name of actor2 column}

\item{time}{name of time column if relevant}
}
\value{
a vector of symmetric ids that can be added back to the data.frame object
}
\description{
This function creates a symmetric id
for a given pair of actors
}
\author{
Shahryar Minhas
}
