% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacency_list.R
\name{get_adjacency_list_old}
\alias{get_adjacency_list_old}
\title{Longitudinal variants of \code{get_adjacency()} which returns a list of adjacency matrices}
\usage{
get_adjacency_list_old(
  dyad_data,
  actor1 = NULL,
  actor2 = NULL,
  time = NULL,
  symmetric = TRUE,
  mode = "unipartite",
  weight = NULL,
  sum_dyads = FALSE,
  actor_time_uniform = FALSE,
  actor_pds = NULL,
  diag_to_NA = TRUE,
  missing_to_zero = TRUE
)
}
\arguments{
\item{dyad_data}{a dyadic dataframe (or a tibble)}

\item{actor1}{character: name of the actor 1 variable in dyad_data}

\item{actor2}{character: name of the actor 2 variable in dyad_data}

\item{time}{character: name of the time variable in dyad_data, the values of the time variable itself should be numeric}

\item{symmetric}{logical: whether ties are symmetric, default is TRUE}

\item{mode}{character: whether the network is unipartite or bipartite, default is unipartite}

\item{weight}{character: name of the weighted edges variable, default is NULL}

\item{sum_dyads}{logical: whether to sum up the \code{weight} value when there exists repeating dyads within time periods}

\item{actor_time_uniform}{logical: whether to assume
actors are the same across the full time series observed in the data
TRUE means that actors are the same across the full time
series observed in the data
FALSE means that actors come in and out of the observed data and
their "existence" should be determined by the data, meaning that
their first year of existence will be determined by the time point
of their first event and their last year of existence by the
time point of their last event}

\item{actor_pds}{a data.frame indicating start and end time point for every
actor, this can be created manually (see example) or using \code{get_actor_time_info.R},
if provided then choice of \code{actor_time_uniform} is irrelevant.}

\item{diag_to_NA}{logical: whether diagonals should be set to NA, default is TRUE}

\item{missing_to_zero}{logical: whether missing values should be set to zero, default is TRUE}
}
\value{
a list of adjacency matrices of class netify
}
\description{
This function takes in a dyadic dataset
and outputs a list of adjacency matrices in which the actor composition in the network can vary over time.
}
\author{
Cassy Dorff, Ha Eun Choi, Shahryar Minhas
}
