% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_helpers.R
\name{get_attribs}
\alias{get_attribs}
\title{Helper function for layer_netify to extract attributes
from listed netlet objects}
\usage{
get_attribs(a_list, attrib, list_format = FALSE, get_unique = FALSE)
}
\arguments{
\item{a_list}{named list object}

\item{attrib}{character string of attribute to extract}

\item{list_format}{logical, if TRUE return list of attributes}

\item{get_unique}{logical, if TRUE return unique values of attribute}
}
\value{
attribute values
}
\description{
Helper function for layer_netify to extract attributes
from listed netlet objects
}
