% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_matrix}
\alias{get_matrix}
\title{This function fills in an adjacency matrix based on actors and data}
\usage{
get_matrix(
  n_rows,
  n_cols,
  actors_rows,
  actors_cols,
  matRowIndices,
  matColIndices,
  value,
  symmetric
)
}
\arguments{
\item{n_rows}{integer specifying number of row actors}

\item{n_cols}{integer specifying number of column actors}

\item{actors_rows}{character vector of row actors}

\item{actors_cols}{character vector of column actors}

\item{matRowIndices}{numeric vector indicating positions of senders in data}

\item{matColIndices}{numeric vector indicating positions of receivers in data}

\item{value}{numeric vector of values to fill in cross-sections of adjacency matrices}

\item{symmetric}{logical indicating whether ties are symmetric}
}
\value{
an adjacency matrix
}
\description{
This function fills in an adjacency matrix based on actors and data
}
\author{
Shahryar Minhas, Ha Eun Choi
}
