% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_matrix_batch}
\alias{get_matrix_batch}
\title{Batch processing version for multiple time periods}
\usage{
get_matrix_batch(
  n_rows_vec,
  n_cols_vec,
  actors_rows_list,
  actors_cols_list,
  matRowIndices_list,
  matColIndices_list,
  value_list,
  symmetric,
  missing_to_zero = TRUE,
  diag_to_NA = TRUE
)
}
\arguments{
\item{n_rows_vec}{integer vector of number of rows for each matrix}

\item{n_cols_vec}{integer vector of number of columns for each matrix}

\item{actors_rows_list}{list of character vectors for row actors}

\item{actors_cols_list}{list of character vectors for column actors}

\item{matRowIndices_list}{list of integer vectors for row indices}

\item{matColIndices_list}{list of integer vectors for column indices}

\item{value_list}{list of numeric vectors for values}

\item{symmetric}{logical indicating whether ties are symmetric}

\item{missing_to_zero}{logical indicating whether missing values should be set to zero}

\item{diag_to_NA}{logical indicating whether diagonal should be set to NA}
}
\value{
a list of adjacency matrices
}
\description{
Batch processing version for multiple time periods
}
\author{
Shahryar Minhas, Ha Eun Choi
}
