% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_matrix_integer}
\alias{get_matrix_integer}
\title{This function fills in an adjacency matrix based on actors and data (INTEGER VERSION)}
\usage{
get_matrix_integer(
  n_rows,
  n_cols,
  actors_rows,
  actors_cols,
  matRowIndices,
  matColIndices,
  value,
  symmetric,
  missing_to_zero = TRUE,
  diag_to_NA = TRUE
)
}
\arguments{
\item{n_rows}{integer specifying number of row actors}

\item{n_cols}{integer specifying number of column actors}

\item{actors_rows}{character vector of row actors}

\item{actors_cols}{character vector of column actors}

\item{matRowIndices}{numeric vector indicating positions of senders in data}

\item{matColIndices}{numeric vector indicating positions of receivers in data}

\item{value}{integer vector of values to fill in cross-sections of adjacency matrices}

\item{symmetric}{logical indicating whether ties are symmetric}

\item{missing_to_zero}{logical indicating whether missing values should be set to zero}

\item{diag_to_NA}{logical indicating whether diagonal should be set to NA}
}
\value{
an integer adjacency matrix
}
\description{
This function fills in an adjacency matrix based on actors and data (INTEGER VERSION)
}
\author{
Shahryar Minhas, Ha Eun Choi
}
