% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify.R
\name{get_netify}
\alias{get_netify}
\title{Constructs a netlet object}
\usage{
get_netify(
  data = matrix(data = NA_real_, nrow = 1, ncol = 1),
  netify_type = NULL,
  actor_pds = list(NULL),
  actor_time_uniform = TRUE,
  weight = list(NULL),
  symmetric = TRUE,
  isolates = FALSE,
  diag_to_NA = TRUE,
  missing_to_zero = TRUE,
  nodal_data = list(NULL),
  dyad_data = list(NULL),
  graph_data = list(NULL),
  ...
)
}
\arguments{
\item{data}{data object}

\item{netify_type}{type of netlet object}

\item{actor_pds}{actor year data object}

\item{actor_time_uniform}{logical for whether time is uniform}

\item{weight}{weighted or not}

\item{symmetric}{logical for whether data is symmetric}

\item{isolates}{logical for what to do with isolates}

\item{diag_to_NA}{logical for how to treat diagonal values}

\item{missing_to_zero}{logical for how to treat missing values}

\item{nodal_data}{nodal data attribute}

\item{dyad_data}{dyad data attribute}

\item{graph_data}{graph data attribute}

\item{...}{additional parameters}
}
\value{
netlet object
}
\description{
\code{get_netify} is a helper function for users to create a netify object.
Not visible to users.
}
