% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_raw.R
\name{get_raw}
\alias{get_raw}
\title{get raw network data without netify attributes}
\usage{
get_raw(netlet)
}
\arguments{
\item{netlet}{a netify object}
}
\value{
a matrix or list object with netify attributes removed. the structure
returned depends on the type of netify object:
\itemize{
\item cross-sectional networks: returns a matrix
\item longitudinal array networks: returns a matrix
\item longitudinal list networks: returns a list
}
}
\description{
\code{get_raw} extracts the underlying network data structure (matrix or list)
from a netify object, removing all netify-specific attributes. this is useful
when you need to work with the base r data structures or pass the network
data to functions that don't recognize netify objects.
}
\author{
shahryar minhas
}
