% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_params.R
\name{gg_params}
\alias{gg_params}
\title{Prepare graphical parameters for ggplot2 based on specified plot arguments.}
\usage{
gg_params(plot_args)
}
\arguments{
\item{plot_args}{A list containing user-defined specifications for various graphical parameters
such as color, size, alpha, and others for different components of the plot.}
}
\value{
A list containing organized graphical parameters for nodes, text, labels, and edges.
Each component has sub-lists for static parameters and variable parameters.
}
\description{
This function organizes and formats static and variable-dependent aesthetic parameters
for nodes, text labels, and edges to be used in ggplot2 visualizations. It adjusts parameters
based on the existence and non-null status of the corresponding entries in \code{plot_args}.
}
\details{
The function splits the graphical parameters into static and dynamic (variable-dependent) categories.
Static parameters are directly applied when the corresponding variable-dependent parameter is not set.
If a variable-dependent parameter is set, the static parameter is excluded to allow dynamic mapping in ggplot.

Dynamic or variable-dependent parameters are set based on a mapping from data columns specified in \code{plot_args}.
The function checks for non-null entries in \code{plot_args} and assigns these to the appropriate aesthetic
if the entry exists. For example, if \code{plot_args$point_color_var} is not null, it will create a dynamic color
mapping for nodes and remove any static color setting.
}
\author{
Cassy Dorff, Shahryar Minhas
}
