% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_label}
\alias{ggplot_add.netify_label}
\title{add netify_label to ggplot}
\usage{
\method{ggplot_add}{netify_label}(object, plot, ...)
}
\arguments{
\item{object}{a netify_label object created by \code{\link{netify_label}}}

\item{plot}{a ggplot object to which the label layer will be added}

\item{...}{additional arguments passed by ggplot2 (used internally)}
}
\value{
a ggplot object with the label layer added
}
\description{
s3 method to add netify_label objects to ggplot objects. this method is called
automatically when using the + operator with a netify_label object.
}
\author{
cassy dorff, shahryar minhas
}
