% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_labels}
\alias{ggplot_add.netify_labels}
\title{add netify scale labels to ggplot}
\usage{
\method{ggplot_add}{netify_labels}(object, plot, ...)
}
\arguments{
\item{object}{a netify_labels object created by \code{\link{netify_scale_labels}}}

\item{plot}{a ggplot object to which the labels will be added}

\item{...}{additional arguments passed by ggplot2 (used internally)}
}
\value{
a ggplot object with updated scale labels
}
\description{
s3 method to add netify_labels objects to ggplot objects. this method updates
the labels of existing scales based on the specifications in the netify_labels
object.
}
\author{
cassy dorff, shahryar minhas
}
