% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_node}
\alias{ggplot_add.netify_node}
\title{add netify_node to ggplot}
\usage{
\method{ggplot_add}{netify_node}(object, plot, ...)
}
\arguments{
\item{object}{a netify_node object created by \code{\link{netify_node}}}

\item{plot}{a ggplot object to which the node layer will be added}

\item{...}{additional arguments passed by ggplot2 (used internally)}
}
\value{
a ggplot object with the node layer added
}
\description{
s3 method to add netify_node objects to ggplot objects. this method is called
automatically when using the + operator with a netify_node object.
}
\author{
cassy dorff, shahryar minhas
}
