% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_scale_reset}
\alias{ggplot_add.netify_scale_reset}
\title{add scale resets to ggplot}
\usage{
\method{ggplot_add}{netify_scale_reset}(object, plot, ...)
}
\arguments{
\item{object}{a netify_scale_reset object created by \code{\link{reset_scales}}}

\item{plot}{a ggplot object to which scale resets will be added}

\item{...}{additional arguments passed by ggplot2 (used internally)}
}
\value{
a ggplot object with scale resets applied
}
\description{
s3 method to add scale reset objects to ggplot objects. this method resets
color, fill, alpha, and size scales using the ggnewscale package.
}
\author{
cassy dorff, shahryar minhas
}
