% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_text}
\alias{ggplot_add.netify_text}
\title{add netify_text to ggplot}
\usage{
\method{ggplot_add}{netify_text}(object, plot, ...)
}
\arguments{
\item{object}{a netify_text object created by \code{\link{netify_text}}}

\item{plot}{a ggplot object to which the text layer will be added}

\item{...}{additional arguments passed by ggplot2 (used internally)}
}
\value{
a ggplot object with the text layer added
}
\description{
s3 method to add netify_text objects to ggplot objects. this method is called
automatically when using the + operator with a netify_text object.
}
\author{
cassy dorff, shahryar minhas
}
