% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_text_repel}
\alias{ggplot_add.netify_text_repel}
\title{add netify_text_repel to ggplot}
\usage{
\method{ggplot_add}{netify_text_repel}(object, plot, ...)
}
\arguments{
\item{object}{a netify_text_repel object created by \code{\link{netify_text_repel}}}

\item{plot}{a ggplot object to which the text_repel layer will be added}

\item{...}{additional arguments passed by ggplot2 (used internally)}
}
\value{
a ggplot object with the text_repel layer added
}
\description{
s3 method to add netify_text_repel objects to ggplot objects. this method is called
automatically when using the + operator with a netify_text_repel object.
}
\author{
cassy dorff, shahryar minhas
}
