% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{identical_recursive}
\alias{identical_recursive}
\title{identical_recursive}
\usage{
identical_recursive(...)
}
\arguments{
\item{...}{objects to check}
}
\value{
logical indicating whether objects are identical
}
\description{
Recursively check if two or more objects are identical
}
\author{
Shahryar Minhas
}
