% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{list_network_styles}
\alias{list_network_styles}
\title{list available network styles}
\usage{
list_network_styles()
}
\value{
a data frame with style names and descriptions
}
\description{
shows all available preset network styles with descriptions
}
\author{
cassy dorff, shahryar minhas
}
