% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{list available color palettes}
\usage{
list_palettes()
}
\value{
character vector of available palette names
}
\description{
list available color palettes
}
\author{
cassy dorff, shahryar minhas
}
