% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_to_array}
\alias{list_to_array}
\title{list_to_array}
\usage{
list_to_array(list_of_mats)
}
\arguments{
\item{list_of_mats}{list object}
}
\value{
three dimensional array
}
\description{
This function converts a list of matrices
into a three dimensional array
}
\author{
Shahryar Minhas
}
