% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{logical_check}
\alias{logical_check}
\title{logical_check}
\usage{
logical_check(
  sum_dyads,
  symmetric,
  diag_to_NA,
  missing_to_zero,
  actor_time_uniform = NULL
)
}
\arguments{
\item{sum_dyads}{user supplied input}

\item{symmetric}{user supplied input}

\item{diag_to_NA}{user supplied input}

\item{missing_to_zero}{user supplied input}

\item{actor_time_uniform}{optional user supplied input}
}
\value{
returns a NULL object but stops the process
if there is an error detected
}
\description{
Checks to make sure user has correctly inputted logicals
for select inputs
}
\author{
Ha Eun Choi, Shahryar Minhas
}
