% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlet_to_array.R
\name{longit_dv_to_arr}
\alias{longit_dv_to_arr}
\title{Convert list of dependent variable(s) into an array}
\usage{
longit_dv_to_arr(netlet)
}
\arguments{
\item{netlet}{netify object}
}
\value{
An array object of dimensions nr x nc x t,
where nr is the number of row actors, nc is the number
of column actors, and t is the number of time periods.
}
\description{
Mainly for use in going from a list netlet to an array netlet
for bridging netlet objects to analysis with packages that
expect matrix/array inputs
}
