% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlet_to_array.R
\name{longit_nodal_to_arr}
\alias{longit_nodal_to_arr}
\title{Convert nodal attribute of netlet into an array}
\usage{
longit_nodal_to_arr(netlet)
}
\arguments{
\item{netlet}{netify object}
}
\value{
a list object of length two, one for the row actors and
another for the column actors. Each element in the list is an
array of dimensions n x pn x t, where n is the number of actors,
pn is the number of nodal variables, and t is the number of time
periods. The rownames of the array are the actors, and the
colnames are the nodal variables.
}
\description{
Mainly for use in going from a list netlet to an array netlet
for bridging netlet objects to analysis with packages that
expect matrix/array inputs
}
