% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\docType{data}
\name{myanmar}
\alias{myanmar}
\title{event data slice from ucdp on myanmar}
\format{
a data frame with dyadic event observations.
}
\source{
ucdp georeferenced event dataset, version 23.1.
\url{https://ucdp.uu.se/downloads/}
}
\usage{
data(myanmar)
}
\description{
event data from ucdp for myanmar
}
\examples{
data(myanmar)
myanmar[1:3, ]

}
\references{
davies, s., pettersson, t., and oberg, m. (2023).
organized violence 1989-2022, and the return of conflict between states.
journal of peace research, 60(4), 691-708.
\doi{10.1177/00223433231185169}.

sundberg, r. and melander, e. (2013). introducing the ucdp
georeferenced event dataset. journal of peace research, 50(4), 523-532.
\doi{10.1177/0022343313484347}.
}
\author{
cassy dorff, shahryar minhas
}
\keyword{datasets}
