% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_label.R
\name{net_label}
\alias{net_label}
\title{Plotting method for node label in 'netify' visualizations}
\usage{
net_label(x, ...)
}
\arguments{
\item{x}{A 'netify' object, which contains the network data structured for analysis and visualization.}

\item{...}{Additional arguments, which can include but are not limited to:
- \code{point_layout}: Optional, user-provided node layout; if not provided, layout will be generated based on \code{layout} parameter.
- \code{layout}: Specifies the layout algorithm from 'igraph' to position the nodes if \code{point_layout} is not provided.
- \code{remove_isolates}: Logical; if TRUE, isolates will be removed from the plot. Default is TRUE.
- \code{static_actor_positions} Logical indicating whether to use static positions for actors.
Useful in longitudinal studies where node positions should remain consistent over time.
If TRUE, the layout by default is calculated based on a collapsed adjacency matrix across
all time points. Users can also specify a specific time point to use as the static layout by
setting \code{which_static} to the desired time point.
- \code{add_edges}: Logical; if TRUE, edges will be added to the plot. Default is TRUE.
- \code{curve_edges}: Logical; if TRUE, edges will be curved. Default is FALSE.
- \code{add_points}: Logical; if TRUE, points (nodes) will be plotted. Default is TRUE.
- \code{add_text}: Logical; if TRUE, text annotations will be added. Default is FALSE.
- \code{add_label}: Logical; if TRUE, labels will be added. Default is FALSE.
- \code{select_text}: A vector of node names to specifically add text to; others will not have text.
- \code{select_label}: A vector of node names to specifically add labels to; others will not have labels.
- \code{point_size}: A fixed size for all points, equivalent to \code{geom_point(size = point_size)} in \code{ggplot2}.
- \code{point_size_var}: A variable from the node data to dynamically size points, equivalent to \code{geom_point(aes(size = point_size_var))} in \code{ggplot2}.
These arguments control various aspects of the plot's appearance and functionality.}
}
\value{
A 'ggplot' plot object that can be further modified or printed.
}
\description{
This function provides a comprehensive tool to visualize 'netify' objects through various graphical representations including points, edges, texts, and labels. It leverages the capabilities of the 'igraph' and 'ggplot' packages to create customizable network visualizations.
}
\author{
Cassy Dorff, Shahryar Minhas
}
