% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify.R
\name{netify_check}
\alias{netify_check}
\title{netify_check}
\usage{
netify_check(netlet)
}
\arguments{
\item{netlet}{user inputted object to check}
}
\value{
NULL object but stops the process if there
is an error detected
}
\description{
Checks to make sure that object is of class netify
and stops process if not
}
\author{
Ha Eun Choi, Colin Henry, Shahryar Minhas
}
