% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{netify_text_repel}
\alias{netify_text_repel}
\title{extract text_repel layer from netify plot components}
\usage{
netify_text_repel(comp)
}
\arguments{
\item{comp}{a netify_plot_components object returned from
\code{plot(..., return_components = TRUE)}}
}
\value{
a custom object of class "netify_text_repel" that can be added to a ggplot
object using the + operator. the object contains the text_repel layer with all
its aesthetic mappings and data.
}
\description{
extracts the text_repel layer from a netify plot components object. text repel labels
display actor names with automatic repositioning to avoid overlaps, making them more
readable in dense networks.
}
\examples{
\dontrun{
# create a netify object
net <- netify(my_data, actor1 = "from", actor2 = "to")

# get plot components with text_repel
comp <- plot(net, add_text_repel = TRUE, return_components = TRUE)

# build custom plot with repelled text
library(ggplot2)
ggplot() +
    netify_text_repel(comp)
}

}
\seealso{
\code{\link{plot.netify}}, \code{\link{netify_text}},
\code{\link{assemble_netify_plot}}
}
\author{
cassy dorff, shahryar minhas
}
