% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_network.R
\name{netify_to_network}
\alias{netify_to_network}
\title{netify_to_network}
\usage{
netify_to_network(netlet)
}
\arguments{
\item{netlet}{netify object}
}
\value{
network object
}
\description{
Convert netify object to network object
}
\author{
Shahryar Minhas
}
