% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_stats.R
\name{plot.summary.netify}
\alias{plot.summary.netify}
\title{plot graph-level summary statistics}
\usage{
\method{plot}{summary.netify}(x, ...)
}
\arguments{
\item{x}{a graph-level summary returned by \code{summary()} for a netify object.}

\item{...}{additional arguments passed to \code{\link[=plot_graph_stats]{plot_graph_stats()}}.}
}
\value{
a ggplot object displaying graph-level summary statistics.
}
\description{
plot graph-level summary statistics
}
