% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networkStats.R
\name{plot_networkStats}
\alias{plot_networkStats}
\title{Plotting function for network level statistics for netify objects}
\usage{
plot_networkStats(summary_df, longitudinal = TRUE, net_stat = net_stat)
}
\arguments{
\item{summary_df}{dataframe produced by summary}

\item{longitudinal}{logical: TRUE if longitudinal network, default is TRUE}

\item{net_stat}{character: network-level network statistics, default is "dens"}
}
\value{
a graph object summarizing network level stats
}
\description{
\code{plot_networkStats} takes in the output from \code{summary.netify} function
and outputs a graph with network level statistics.
}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
