% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{plot_with_style}
\alias{plot_with_style}
\title{Apply style to netify plot}
\usage{
plot_with_style(netlet, style_fun, ...)
}
\arguments{
\item{netlet}{a netify object}

\item{style_fun}{a style function (e.g., style_rose)}

\item{...}{additional plot parameters}
}
\value{
a ggplot object
}
\description{
helper function to properly apply styles to netify plots
}
\author{
cassy dorff, shahryar minhas
}
