% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_network.R
\name{prep_for_statnet}
\alias{prep_for_statnet}
\title{Convert conflictNet object to a statnet network object}
\usage{
prep_for_statnet(netlet)
}
\arguments{
\item{netlet}{An R object}
}
\value{
statnet network object
}
\description{
Convert conflictNet object to a statnet network object
}
\examples{

# load data
data(icews)

# cross-sectional case
icews_10 <- icews[icews$year==2010,]

# create netify object
dvars = c( 'matlCoop', 'verbConf', 'matlConf' )
nvars = c( 'i_polity2','i_log_gdp', 'i_log_pop' )
verbCoop_net = netify( 
  icews_10,
  actor1='i', actor2='j', 
  symmetric=FALSE, 
  weight='verbCoop',
  dyad_vars = dvars,
  dyad_vars_symmetric=rep(FALSE, length(dvars)),
  nodal_vars = nvars )

# convert to a statnet network object
ntwk <- prep_for_statnet(verbCoop_net)
ntwk

# longitudinal case
verbCoop_longit_net = netify(
  icews,
  actor1='i', actor2='j', time='year',
  symmetric=FALSE, 
  weight='verbCoop',
  dyad_vars = dvars,
  dyad_vars_symmetric=rep(FALSE, length(dvars)),
  nodal_vars = nvars )

# convert to a statnet network object
ntwk_longit <- prep_for_statnet(verbCoop_longit_net)

# output in the longitudinal case is 
# a list of statnet network objects
class(ntwk_longit)
names(ntwk_longit)
ntwk_longit[['2002']]

}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
