% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_raw.R
\name{raw_list}
\alias{raw_list}
\title{raw_list}
\usage{
raw_list(netlet)
}
\arguments{
\item{netlet}{a netify object of type 'longit_list'}
}
\value{
an object of class list with netify
attributes stripped
}
\description{
This function takes in a netify object of type
cross_sec and returns a regular matrix
}
\author{
Shahryar Minhas
}
