% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_raw.R
\name{raw_matrix}
\alias{raw_matrix}
\title{raw_matrix}
\usage{
raw_matrix(netlet)
}
\arguments{
\item{netlet}{a netify object of type
cross_sec' or 'longit_array'}
}
\value{
an object of class matrix, array with netify
attributes stripped
}
\description{
This function takes in a netify object of type
cross_sec or longit_array and returns an R matrix/array
}
\author{
Shahryar Minhas
}
