% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_helpers.R
\name{reduce_combine_dyad_attr}
\alias{reduce_combine_dyad_attr}
\title{Reduce and combine multiple dyadic attributes of netify objects
into a single dyadic attribute. Mainly for use within the
layer_netlet function}
\usage{
reduce_combine_dyad_attr(attribs_list, msrmnts_list, netlet_type)
}
\arguments{
\item{attribs_list}{list of attributes from each netlet}

\item{msrmnts_list}{list of msrmnts from each netlet}

\item{netlet_type}{character string of netlet type}
}
\value{
dyad attribute data.frame
}
\description{
Reduce and combine multiple dyadic attributes of netify objects
into a single dyadic attribute. Mainly for use within the
layer_netlet function
}
