% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{repeat_dyads_check}
\alias{repeat_dyads_check}
\title{repeat_dyads_check}
\usage{
repeat_dyads_check(dyad_data, actor1, actor2, time = NULL)
}
\arguments{
\item{dyad_data}{user inputted data.frame object}

\item{actor1}{character value denoting actor1 in the data.frame object}

\item{actor2}{character value denoting actor2 in the data.frame object}

\item{time}{optoinal character value denoting time in the data.frame object}
}
\value{
returns a count of the number of repeating dyads in the
data.frame object
}
\description{
Check whether dyadic observations are
repeating in the data.frame object
}
\author{
Ha Eun Choi, Shahryar Minhas
}
