% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repeat_dyads_check_cpp}
\alias{repeat_dyads_check_cpp}
\title{Determine number of repeating dyad-time obs}
\usage{
repeat_dyads_check_cpp(actor1, actor2, time)
}
\arguments{
\item{actor1}{character vector for actor1}

\item{actor2}{character vector for actor2}

\item{time}{numeric vector for time}
}
\value{
an integer count of number of repeating dyads
}
\description{
Determine number of repeating dyad-time obs
}
\author{
Shahryar Minhas
}
