% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_helpers.R
\name{set_layer_labels}
\alias{set_layer_labels}
\title{Helper function for layer_netify to set layer labels}
\usage{
set_layer_labels(netlet_list, layer_labels)
}
\arguments{
\item{netlet_list}{list of netlet objects}

\item{layer_labels}{character vector of layer labels}
}
\value{
character vector of layer labels
}
\description{
Helper function for layer_netify to set layer labels
}
