% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_string}
\alias{split_string}
\title{Break string into list of strings by some fixed character
and then extract the desired values around that fixed
character}
\usage{
split_string(string_to_split, break_by, to_extract, fixed = TRUE)
}
\arguments{
\item{string_to_split}{character: string to be split}

\item{break_by}{character: character to break string by}

\item{to_extract}{integer: index of the string to be extracted}

\item{fixed}{If \code{TRUE} match exactly, otherwise use regular expressions}
}
\value{
a character vector of the extracted strings
}
\description{
Break string into list of strings by some fixed character
and then extract the desired values around that fixed
character
}
\author{
Shahryar Minhas
}
