% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_538}
\alias{style_538}
\title{FiveThirtyEight-inspired network style}
\usage{
style_538()
}
\value{
List of plot arguments
}
\description{
A network style inspired by FiveThirtyEight's data visualization
style with gray background and bright accents
}
