% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_black_yellow}
\alias{style_black_yellow}
\title{black and yellow network style}
\usage{
style_black_yellow()
}
\value{
list of plot arguments
}
\description{
a network style with dark green, black, and yellow colors.
}
\author{
cassy dorff, shahryar minhas
}
