% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_bronze_block}
\alias{style_bronze_block}
\title{bronze block network style}
\usage{
style_bronze_block()
}
\value{
list of plot arguments
}
\description{
a network style with bold, substantial visuals.
}
\author{
cassy dorff, shahryar minhas
}
