% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_budapest}
\alias{style_budapest}
\title{Wes Anderson-inspired network style (Grand Budapest Hotel)}
\usage{
style_budapest()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the pastel color palette of
Wes Anderson's Grand Budapest Hotel
}
