% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_crimson_silver}
\alias{style_crimson_silver}
\title{crimson and silver network style}
\usage{
style_crimson_silver()
}
\value{
list of plot arguments
}
\description{
a network style with sharp contrasts and sleek design.
}
\author{
cassy dorff, shahryar minhas
}
