% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_cyberpunk}
\alias{style_cyberpunk}
\title{cyberpunk-inspired network style}
\usage{
style_cyberpunk()
}
\value{
list of plot arguments
}
\description{
a futuristic network style with neon colors on dark background
}
\author{
cassy dorff, shahryar minhas
}
