% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_dark2}
\alias{style_dark2}
\title{colorbrewer dark2-based network style}
\usage{
style_dark2()
}
\value{
list of plot arguments
}
\description{
a network style using colorbrewer's dark2 palette
for high contrast and accessibility
}
\author{
cassy dorff, shahryar minhas
}
