% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_economist}
\alias{style_economist}
\title{Economist-style network style}
\usage{
style_economist()
}
\value{
List of plot arguments
}
\description{
A network style inspired by The Economist's distinctive
visual style with red accents and clean lines
}
