% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_green_gold}
\alias{style_green_gold}
\title{green and gold network style}
\usage{
style_green_gold()
}
\value{
list of plot arguments
}
\description{
a network style with balanced green, gold, and blue colors.
}
\author{
cassy dorff, shahryar minhas
}
