% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_lime_magenta}
\alias{style_lime_magenta}
\title{lime and magenta network style}
\usage{
style_lime_magenta()
}
\value{
list of plot arguments
}
\description{
a network style with bright green and magenta contrast.
}
\author{
cassy dorff, shahryar minhas
}
