% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_nature}
\alias{style_nature}
\title{Nature journal-inspired network style}
\usage{
style_nature()
}
\value{
List of plot arguments
}
\description{
A network style inspired by Nature journal's clean scientific
visualization style
}
