% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_navy_maroon}
\alias{style_navy_maroon}
\title{navy and maroon network style}
\usage{
style_navy_maroon()
}
\value{
list of plot arguments
}
\description{
a network style with navy blue and maroon colors.
}
\author{
cassy dorff, shahryar minhas
}
