% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_orange_teal}
\alias{style_orange_teal}
\title{orange and teal network style}
\usage{
style_orange_teal()
}
\value{
list of plot arguments
}
\description{
a network style with gray background and bright accents.
}
\author{
cassy dorff, shahryar minhas
}
