% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_pastel}
\alias{style_pastel}
\title{pastel rainbow network style}
\usage{
style_pastel()
}
\value{
list of plot arguments
}
\description{
a soft, cheerful network style using pastel colors
}
\author{
cassy dorff, shahryar minhas
}
