% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_pokemon_pikachu}
\alias{style_pokemon_pikachu}
\title{Pokemon-style network style (Pikachu)}
\usage{
style_pokemon_pikachu()
}
\value{
List of plot arguments
}
\description{
A playful style inspired by Pikachu: energetic yellows, bold outlines, and lightning vibes
}
