% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_racing_blue}
\alias{style_racing_blue}
\title{racing blue network style}
\usage{
style_racing_blue()
}
\value{
list of plot arguments
}
\description{
a sleek style with clean lines, blue accents, and strong contrast.
}
\author{
cassy dorff, shahryar minhas
}
