% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_random}
\alias{style_random}
\title{apply random network style}
\usage{
style_random(seed = NULL)
}
\arguments{
\item{seed}{optional seed for reproducibility}
}
\value{
a style list that can be passed to plot()
}
\description{
randomly selects and returns one of the available network styles
}
\author{
cassy dorff, shahryar minhas
}
