% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_red_blue}
\alias{style_red_blue}
\title{red and blue network style}
\usage{
style_red_blue()
}
\value{
list of plot arguments
}
\description{
a network style with red accents and clean lines.
}
\author{
cassy dorff, shahryar minhas
}
