% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_retro80s}
\alias{style_retro80s}
\title{retro 80s-inspired network style}
\usage{
style_retro80s()
}
\value{
list of plot arguments
}
\description{
a network style channeling 1980s design aesthetics
with bold colors and geometric shapes
}
\author{
cassy dorff, shahryar minhas
}
