% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_rose}
\alias{style_rose}
\title{rose network style}
\usage{
style_rose()
}
\value{
list of plot arguments
}
\description{
a network style with soft rose, purple, and gold colors.
}
\author{
cassy dorff, shahryar minhas
}
