% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_scientific_blue}
\alias{style_scientific_blue}
\title{scientific blue network style}
\usage{
style_scientific_blue()
}
\value{
list of plot arguments
}
\description{
a network style with clean blue and orange colors.
}
\author{
cassy dorff, shahryar minhas
}
