% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_secretariat}
\alias{style_secretariat}
\title{Secretariat-style network style}
\usage{
style_secretariat()
}
\value{
List of plot arguments
}
\description{
A powerful, sleek style inspired by Secretariat - clean lines, racing blues, and commanding contrast.
}
