% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_slate_silver}
\alias{style_slate_silver}
\title{slate and silver network style}
\usage{
style_slate_silver()
}
\value{
list of plot arguments
}
\description{
a network style with steel gray and silver colors.
}
\author{
cassy dorff, shahryar minhas
}
