% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_solarized}
\alias{style_solarized}
\title{solarized-inspired network style}
\usage{
style_solarized()
}
\value{
list of plot arguments
}
\description{
a network style based on the popular solarized color scheme,
optimized for reduced eye strain
}
\author{
cassy dorff, shahryar minhas
}
