% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_stata}
\alias{style_stata}
\title{Stata-inspired network style}
\usage{
style_stata()
}
\value{
List of plot arguments
}
\description{
A network style that mimics Stata's default graph colors
with navy blue and maroon
}
