% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_sunburst}
\alias{style_sunburst}
\title{sunburst network style}
\usage{
style_sunburst()
}
\value{
list of plot arguments
}
\description{
a playful style with energetic yellows and bold outlines.
}
\author{
cassy dorff, shahryar minhas
}
