% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_temporal_network}
\alias{style_temporal_network}
\title{preset style for temporal/longitudinal networks}
\usage{
style_temporal_network()
}
\value{
list of plot arguments
}
\description{
a complete visual style optimized for displaying networks over time
}
\author{
cassy dorff, shahryar minhas
}
