% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_tufte}
\alias{style_tufte}
\title{minimal tufte-inspired network style}
\usage{
style_tufte()
}
\value{
list of plot arguments
}
\description{
a minimalist network style inspired by edward tufte's
data visualization principles
}
\author{
cassy dorff, shahryar minhas
}
