% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_witcher_feline}
\alias{style_witcher_feline}
\title{Witcher 3 Feline School network style}
\usage{
style_witcher_feline()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the Cat School armor from The Witcher 3.
Light, agile, with emphasis on critical strikes - translates to sharp
contrasts and sleek design.
}
