% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_witcher_griffin}
\alias{style_witcher_griffin}
\title{Witcher 3 Griffin School network style}
\usage{
style_witcher_griffin()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the Griffin School armor from The Witcher 3.
Balanced between signs and swordplay - moderate, magical aesthetics.
}
