% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_witcher_manticore}
\alias{style_witcher_manticore}
\title{Witcher 3 Manticore School network style (bonus)}
\usage{
style_witcher_manticore()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the Manticore School armor from The Witcher 3.
Exotic, alchemical focus - poisonous and mysterious.
}
