% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_witcher_ursine}
\alias{style_witcher_ursine}
\title{Witcher 3 Ursine School network style}
\usage{
style_witcher_ursine()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the Bear School armor from The Witcher 3.
Heavy, tank-like armor - translates to bold, substantial visuals.
}
