% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_witcher_viper}
\alias{style_witcher_viper}
\title{Witcher 3 Viper School network style (bonus)}
\usage{
style_witcher_viper()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the Viper School armor from The Witcher 3.
Lethal, assassination-focused - stealth and poison.
}
