% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_witcher_wolven}
\alias{style_witcher_wolven}
\title{Witcher 3 Wolven School network style}
\usage{
style_witcher_wolven()
}
\value{
List of plot arguments
}
\description{
A network style inspired by the Wolf School armor from The Witcher 3.
The most balanced school - Geralt's home school at Kaer Morhen.
}
