% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_networks_print.R
\name{summary.netify_comparison}
\alias{summary.netify_comparison}
\title{summary method for netify_comparison objects}
\usage{
\method{summary}{netify_comparison}(object, ...)
}
\arguments{
\item{object}{a netify_comparison object from compare_networks()}

\item{...}{additional arguments (currently unused)}
}
\value{
a summary data frame or list depending on comparison type
}
\description{
provides a concise summary of network comparison results.
}
\author{
cassy dorff, shahryar minhas
}
