% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_netify}
\alias{theme_netify}
\title{theme_netify function}
\usage{
theme_netify()
}
\value{
a customized theme object for netify plots.
}
\description{
this function returns a customized theme for netify plots.
it is based on the \code{theme_minimal} function from the \code{ggplot2} package.
it removes axis text and titles from the plot.
}
\author{
cassy dorff, shahryar minhas
}
