% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{time_check}
\alias{time_check}
\title{time_check}
\usage{
time_check(time, dyad_data)
}
\arguments{
\item{time}{user inputted object for weight
and NULL by default}

\item{dyad_data}{data.frame in which time
values are located}
}
\value{
returns a NULL object but stops the process
if there is an error detected
}
\description{
Checks to make sure that the time field
is populated correctly
}
\author{
Ha Eun Choi, Shahryar Minhas
}
