% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{unipartite_actor_check}
\alias{unipartite_actor_check}
\title{Validate unipartite network actor requirements}
\usage{
unipartite_actor_check(actors_1, actors_2)
}
\arguments{
\item{actors_1}{character vector of unique actor1 values}

\item{actors_2}{character vector of unique actor2 values}
}
\value{
invisible(TRUE) if validation passes, otherwise throws error or warning
}
\description{
Validate unipartite network actor requirements
}
\author{
Cassy Dorff, Shahryar Minhas
}
\keyword{internal}
