% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unique_vector}
\alias{unique_vector}
\title{unique_vector}
\usage{
unique_vector(...)
}
\arguments{
\item{...}{vector inputs}
}
\value{
numeric vector
}
\description{
Get unique vector from
multiple vector inputs
}
\author{
Shahryar Minhas
}
