% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_plot_params.R
\name{validate_plot_params}
\alias{validate_plot_params}
\title{Validate plot parameters and warn about common mistakes}
\usage{
validate_plot_params(plot_args, ...)
}
\arguments{
\item{plot_args}{List of all plot arguments}

\item{...}{Additional arguments passed to plot.netify}
}
\value{
NULL (function is called for side effects - warnings)
}
\description{
This function checks the parameter names passed through ... in plot.netify
and warns users about potential typos or deprecated parameter names.
}
\keyword{internal}
