% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{weight_check}
\alias{weight_check}
\title{weight_check}
\usage{
weight_check(weight, dyad_data)
}
\arguments{
\item{weight}{user inputted object for weight
and NULL by default}

\item{dyad_data}{data.frame in which weight
values are located}
}
\value{
returns a NULL object but stops the process
if there is an error detected
}
\description{
Checks to make sure that the weight field
is populated correctly
}
\author{
Ha Eun Choi, Shahryar Minhas
}
