% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{weight_string_label}
\alias{weight_string_label}
\title{weight_string_label}
\usage{
weight_string_label(weight, sum_dyads)
}
\arguments{
\item{weight}{user input for weight}

\item{sum_dyads}{logical user input for sum_dyads}
}
\value{
character string
}
\description{
Create attribute label of for weight
based on user inputs to netify
}
\author{
Shahryar Minhas
}
