% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_across_units.R
\name{agg_across_units}
\alias{agg_across_units}
\title{agg_across_units}
\usage{
agg_across_units(
  dyad_data,
  actor1,
  actor2,
  time = NULL,
  weight,
  symmetric,
  ignore_missing = TRUE
)
}
\arguments{
\item{dyad_data}{user inputted data.frame object}

\item{actor1}{character: actor 1 in the data}

\item{actor2}{character: actor 2 in the data}

\item{time}{character: time in the data}

\item{weight}{character: weight variable in the data}

\item{symmetric}{logical: is the adjacency matrix symmetric}

\item{ignore_missing}{logical: should missing values be ignored}
}
\value{
aggregated data.frame object
}
\description{
Aggregate dyadic dataset
}
\author{
Shahryar Minhas
}
