% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{assemble_netify_plot}
\alias{assemble_netify_plot}
\title{Assemble netify plot from components}
\usage{
assemble_netify_plot(comp)
}
\arguments{
\item{comp}{A netify_plot_components object}
}
\value{
A ggplot object
}
\description{
Assemble netify plot from components
}
