% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actor_time_info.R
\name{get_actor_time_info}
\alias{get_actor_time_info}
\title{Creates a dataframe indicating start and end time point}
\usage{
get_actor_time_info(dyad_data, actor1, actor2, time)
}
\arguments{
\item{dyad_data}{a dyadic dataframe}

\item{actor1}{character: actor 1 in the data}

\item{actor2}{character: actor 2 in the data}

\item{time}{character: time in the data}
}
\value{
a dataframe with three columns:
\enumerate{
\item \item{actor}{a column indicating actors in the data}
\item \item{min_time}{a column indicating the first time
point in which the actor should be considered a part of the network}
\item \item{max_time}{a column indicating the last time point in
which the actor should be considered a part of the network}
}
}
\description{
\code{get_actor_time_info} takes in a longitudinal dyadic dataset and computes when
actors entered and exited the network. Entering will be determined by the
first period the actor had an interaction and exiting by the last period.
}
\examples{

library(peacesciencer)
library(dplyr)
cow_dyads <- create_dyadyears(
  subset_years = c(1980:2001)
  ) \%>\%
  # add mids
  add_cow_mids()
actor_time <- get_actor_time_info(
  dyad_data = cow_dyads, 
  actor1 = 'ccode1', 
  actor2 = 'ccode2', 
  time = 'year'
)

}
\author{
Shahryar Minhas, Ha Eun Choi
}
