% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout.R
\name{get_edge_layout}
\alias{get_edge_layout}
\title{Generate edge layout for netify object}
\usage{
get_edge_layout(netlet, nodes_list)
}
\arguments{
\item{netlet}{A \code{netify} object used to derive the graph and edges.}

\item{nodes}{A matrix of node layouts or list of matrices produced by \code{get_node_layout}.}
}
\value{
A matrix with edge layout information or list of the same
with their start and end coordinates ('x1', 'y1' for the 'from' node and 'x2', 'y2' for the 'to' node).
}
\description{
This function prepares the edge data for visualization by calculating the coordinates
for line segments representing edges based on the node layouts provided by \code{get_node_layout}.
It is specifically tailored for use with \code{netify} objects and their corresponding layout data.
}
\author{
Cassy Dorff, Shahryar Minhas
}
