% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_raw.R
\name{get_raw}
\alias{get_raw}
\title{Get raw network data without netify attributes}
\usage{
get_raw(netlet)
}
\arguments{
\item{netlet}{A netify object}
}
\value{
A matrix or list object with netify attributes removed. The structure
returned depends on the type of netify object:
\itemize{
\item Cross-sectional networks: returns a matrix
\item Longitudinal array networks: returns a matrix
\item Longitudinal list networks: returns a list
}
}
\description{
\code{get_raw} extracts the underlying network data structure (matrix or list)
from a netify object, removing all netify-specific attributes. This is useful
when you need to work with the base R data structures or pass the network
data to functions that don't recognize netify objects.
}
\author{
Shahryar Minhas
}
