% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_raw.R
\name{get_raw}
\alias{get_raw}
\title{get_raw}
\usage{
get_raw(netlet)
}
\arguments{
\item{netlet}{a netify object}
}
\value{
a list or matrix object with netify attributes stripped
}
\description{
This function serves as a wrapper for
raw_matrix and raw_list, which serve to
strip netify related attributes from lists
and matrix obejcts
}
\author{
Shahryar Minhas
}
