% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_edge}
\alias{ggplot_add.netify_edge}
\title{Method to add netify_edge to ggplot}
\usage{
\method{ggplot_add}{netify_edge}(object, plot, object_name)
}
\description{
Method to add netify_edge to ggplot
}
