% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_label}
\alias{ggplot_add.netify_label}
\title{Method to add netify_label to ggplot}
\usage{
\method{ggplot_add}{netify_label}(object, plot, object_name)
}
\description{
Method to add netify_label to ggplot
}
