% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_label}
\alias{ggplot_add.netify_label}
\title{Add netify_label to ggplot}
\usage{
\method{ggplot_add}{netify_label}(object, plot, ...)
}
\arguments{
\item{object}{A netify_label object created by \code{\link{netify_label}}}

\item{plot}{A ggplot object to which the label layer will be added}

\item{...}{Additional arguments passed by ggplot2 (used internally)}
}
\value{
A ggplot object with the label layer added
}
\description{
S3 method to add netify_label objects to ggplot objects. This method is called
automatically when using the + operator with a netify_label object.
}
