% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_labels}
\alias{ggplot_add.netify_labels}
\title{Add netify scale labels to ggplot}
\usage{
\method{ggplot_add}{netify_labels}(object, plot, object_name)
}
\arguments{
\item{object}{A netify_labels object created by \code{\link{netify_scale_labels}}}

\item{plot}{A ggplot object to which the labels will be added}

\item{object_name}{The name of the object being added (used internally by ggplot2)}
}
\value{
A ggplot object with updated scale labels
}
\description{
S3 method to add netify_labels objects to ggplot objects. This method updates
the labels of existing scales based on the specifications in the netify_labels
object.
}
