% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_node}
\alias{ggplot_add.netify_node}
\title{Method to add netify_node to ggplot}
\usage{
\method{ggplot_add}{netify_node}(object, plot, object_name)
}
\description{
Method to add netify_node to ggplot
}
