% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_scale_reset}
\alias{ggplot_add.netify_scale_reset}
\title{Method to add scale resets to ggplot}
\usage{
\method{ggplot_add}{netify_scale_reset}(object, plot, object_name)
}
\description{
Method to add scale resets to ggplot
}
