% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_text}
\alias{ggplot_add.netify_text}
\title{Method to add netify_text to ggplot}
\usage{
\method{ggplot_add}{netify_text}(object, plot, object_name)
}
\description{
Method to add netify_text to ggplot
}
