% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.netify_text_repel}
\alias{ggplot_add.netify_text_repel}
\title{Add netify_text_repel to ggplot}
\usage{
\method{ggplot_add}{netify_text_repel}(object, plot, ...)
}
\arguments{
\item{object}{A netify_text_repel object created by \code{\link{netify_text_repel}}}

\item{plot}{A ggplot object to which the text_repel layer will be added}

\item{...}{Additional arguments passed by ggplot2 (used internally)}
}
\value{
A ggplot object with the text_repel layer added
}
\description{
S3 method to add netify_text_repel objects to ggplot objects. This method is called
automatically when using the + operator with a netify_text_repel object.
}
