% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icews.R
\docType{data}
\name{icews}
\alias{icews}
\title{Event data slice from ICEWS}
\source{
\href{https://dataverse.harvard.edu/dataverse/icews}{icews}
}
\usage{
data(icews)
}
\description{
Event data from ICEWS for select countries from 2002 to 2014,
and additional nodal variables from the World Bank and Polity IV.
}
\examples{
data(icews)
icews[1:3, ]
}
\references{
add icews reference
(\href{https://dataverse.harvard.edu/dataverse/icews}{icews})
}
\keyword{datasets}
