% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify_constructors.R
\name{is_netify}
\alias{is_netify}
\alias{is.netify}
\title{Is this object a netify object?}
\usage{
is_netify(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Logical constant, \code{TRUE} if argument \code{object} is a netify
object
}
\description{
Is this object a netify object?
}
\author{
Colin Henry
}
\keyword{netify}
