% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List available color palettes}
\usage{
list_palettes()
}
\value{
Character vector of available palette names
}
\description{
List available color palettes
}
