% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_ers.R
\name{melt}
\alias{melt}
\title{Melt methods for netify objects}
\usage{
melt(data, ...)
}
\arguments{
\item{data}{An object to be melted (e.g., a netify object)}

\item{...}{Additional arguments passed to methods}
}
\description{
Convert netify matrices/arrays to long format data frames.
These methods provide a consistent interface for melting different
types of netify objects while leveraging C++ for performance.
}
