% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\docType{data}
\name{mexico}
\alias{mexico}
\title{Event data slice from UCDP on Mexico}
\format{
A data frame with dyadic event observations.
}
\source{
\href{https://ucdp.uu.se/}{ucdp}
}
\usage{
data(mexico)
}
\description{
Event data from UCDP for Mexico
}
\examples{
data(mexico)
mexico[1:3, ]
}
\references{
add UCDP reference
(\href{https://ucdp.uu.se/}{ucdp})
}
\keyword{datasets}
