% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\docType{data}
\name{myanmar}
\alias{myanmar}
\title{Event data slice from UCDP on Myanmar}
\source{
\href{https://ucdp.uu.se/}{ucdp}
}
\usage{
data(myanmar)
}
\description{
Event data from UCDP for Myanmar
}
\examples{
data(myanmar)
myanmar[1:3, ]
}
\references{
add UCDP reference
(\href{https://ucdp.uu.se/}{ucdp})
}
\keyword{datasets}
