% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot_data.R
\name{net_plot_data}
\alias{net_plot_data}
\title{Data preparation for plotting 'netify' objects}
\usage{
net_plot_data(netlet, plot_args = list())
}
\arguments{
\item{netlet}{A 'netify' object, which contains the network data structured for analysis and visualization.}

\item{plot_args}{A list of arguments that influence the layout and presentation of the network visualization, including:
- \code{point_layout}: Optional, user-provided node layout; if not provided, layout will be generated based on \code{layout} parameter.
- \code{layout}: Specifies the layout algorithm from 'igraph' to position the nodes if \code{point_layout} is not provided.
- \code{remove_isolates}: Logical; if TRUE, isolates will be removed from the plot. Default is TRUE.
- \code{static_actor_positions}: Logical indicating whether to use static positions for actors.
- \code{add_edges}: Logical; if TRUE, edges will be added to the plot. Default is TRUE.
- \code{curve_edges}: Logical; if TRUE, edges will be curved. Default is FALSE.
- \code{add_points}: Logical; if TRUE, points (nodes) will be plotted. Default is TRUE.
- \code{add_text}: Logical; if TRUE, text annotations will be added. Default is FALSE.
- \code{add_label}: Logical; if TRUE, labels will be added. Default is FALSE.
- \code{select_text}: A vector of node names to specifically add text to; others will not have text.
- \code{select_label}: A vector of node names to specifically add labels to; others will not have labels.}
}
\value{
A list containing \code{plot_args}, \code{ggnet_params}, and \code{net_dfs} which are used for setting up the plot:
- \code{plot_args}: Adjusted plotting arguments including layout and graphical settings.
- \code{ggnet_params}: Graphical parameters organized for nodes, edges, and labels.
- \code{net_dfs}: Data frames for nodes and edges prepared for plotting.
}
\description{
This function prepares the necessary data components for visualizing 'netify' objects.
It processes network attributes and setups up layout and aesthetic parameters for subsequent plotting.
}
\author{
Cassy Dorff, Shahryar Minhas
}
