% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{netify_edge}
\alias{netify_edge}
\title{Extract edges layer from netify plot components}
\usage{
netify_edge(comp)
}
\arguments{
\item{comp}{A netify_plot_components object}
}
\value{
A custom object that can be added to ggplot
}
\description{
Extract edges layer from netify plot components
}
