% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{netify_label}
\alias{netify_label}
\title{Extract label layer from netify plot components}
\usage{
netify_label(comp)
}
\arguments{
\item{comp}{A netify_plot_components object}
}
\value{
A custom object that can be added to ggplot
}
\description{
Extract label layer from netify plot components
}
