% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify_measurements.R
\name{netify_measurements}
\alias{netify_measurements}
\alias{netlet_measurements}
\alias{measurements}
\title{Get measurements of a netify object}
\usage{
netify_measurements(netlet)

netlet_measurements(netlet)

measurements(netlet)
}
\arguments{
\item{netlet}{a netify object}
}
\value{
a list of measurements about the netify object
}
\description{
Get measurements of a netify object
}
\author{
Cassy Dorff, Shahryar Minhas
}
