% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{netify_node}
\alias{netify_node}
\title{Extract nodes layer from netify plot components}
\usage{
netify_node(comp)
}
\arguments{
\item{comp}{A netify_plot_components object}
}
\value{
A custom object that can be added to ggplot
}
\description{
Extract nodes layer from netify plot components
}
