% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_igraph.R
\name{netify_to_igraph}
\alias{netify_to_igraph}
\title{netify_to_igraph}
\usage{
netify_to_igraph(netlet)
}
\arguments{
\item{netlet}{netify object}
}
\value{
igraph object
}
\description{
Convert netify object to igraph object
}
\author{
Shahryar Minhas
}
