% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_network.R
\name{netify_to_statnet}
\alias{netify_to_statnet}
\title{netify_to_statnet}
\usage{
netify_to_statnet(netlet)
}
\arguments{
\item{netlet}{netify object}
}
\value{
statnet network object
}
\description{
Convert netify object to a statnet network object
}
\author{
Shahryar Minhas
}
