% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify_constructors.R
\name{new_netify}
\alias{new_netify}
\title{Constructs a generic netify Object}
\usage{
new_netify(data, ...)
}
\arguments{
\item{data}{A data object (matrix, array, or list of matrices).}

\item{...}{Additional parameters to be stored as attributes on the netify object.}
}
\value{
A netify object with the appropriate attributes for the given data structure.
}
\description{
\code{new_netify} is a low-level constructor for creating new netify objects.
Mostly for internal use, but can be used to create netify objects from
matrices, arrays, or lists of matrices.
}
\author{
Cassy Dorff, Shahryar Minhas
}
