% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify.R
\name{new_netify}
\alias{new_netify}
\title{Constructs a generic netlet Object}
\usage{
new_netify(data, ...)
}
\arguments{
\item{data}{data object}

\item{...}{additional parameters}
}
\value{
a netlet object
}
\description{
\code{new_netify} is a low-level constructor for efficiently creating new netlet objects
Not visible to users.
}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
