% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.netify.R
\name{plot.netify}
\alias{plot.netify}
\title{Plot method for netify objects}
\usage{
\method{plot}{netify}(x, ...)
}
\arguments{
\item{x}{object of class netify, produced by netify()}

\item{...}{Additional arguments passed to geom functions within 'ggraph'.
This can include aesthetics mappings such as alpha, size, and color,
or any other graphical parameters supported by 'geom_edge_link' and 'geom_node_point'.}
}
\value{
ggraph object
}
\description{
\code{plot.netify} takes in a netify object
and produces a network visualization
using the \code{tidygraph} and \code{ggraph} packages.
}
\author{
Cassy Dorff, Shahryar Minhas
}
