% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.netify.R
\name{plot.netify}
\alias{plot.netify}
\title{Plotting method for 'netify' objects}
\usage{
\method{plot}{netify}(x, ...)
}
\arguments{
\item{x}{A 'netify' object, which contains the network data structured for analysis and visualization.}

\item{...}{Additional arguments passed to 'geom' functions (i.e., geom_text, geom_label for nodes and geom_segment for edges) within 'ggplot' for further customization.}

\item{point_layout}{Optional, user-provided node layout; if not provided, layout will be generated based on the \code{layout} parameter.}

\item{layout}{Specifies the layout algorithm from 'igraph' to position the nodes if \code{point_layout} is not provided.}

\item{remove_isolates}{Logical; if TRUE, isolates will be removed from the plot. Default is TRUE.}

\item{add_edges}{Logical; if TRUE, edges will be added to the plot. Default is TRUE.}

\item{curve_edges}{Logical; if TRUE, edges will be curved. Default is FALSE.}

\item{add_points}{Logical; if TRUE, points (nodes) will be plotted. Default is TRUE.}

\item{add_text}{Logical; if TRUE, text annotations will be added. Default is FALSE.}

\item{add_label}{Logical; if TRUE, labels will be added. Default is FALSE.}

\item{select_text}{A vector of node names to specifically add text to; others will not have text.}

\item{select_label}{A vector of node names to specifically add labels to; others will not have labels.}

\item{point_size}{A fixed size for all points, equivalent to \code{geom_point(size = point_size)} in \code{ggplot2}.}

\item{point_size_var}{A variable from the node data to dynamically size points, equivalent to \code{geom_point(aes(size = point_size_var))} in \code{ggplot2}.}
}
\value{
A 'ggplot' plot object that can be further modified or printed.
}
\description{
This function provides a comprehensive tool to visualize 'netify' objects through various graphical representations including points, edges, texts, and labels. It leverages the capabilities of the 'igraph' and 'ggplot' packages to create customizable network visualizations.
}
\details{
\code{plot.netify} creates a network plot with flexible options for customization. It allows the user to specify:
\itemize{
\item Whether or not to include edges and how they are rendered (straight or curved).
\item Whether nodes are displayed as points.
\item Whether to annotate nodes with text or labels.
\item Specific nodes to annotate, allowing selective emphasis within the network.
\item Custom layouts if the automatic placement does not suffice.
Additional customization options like \code{point_size} and \code{point_size_var} allow users to apply typical \code{ggplot2} methods directly to network visualizations, facilitating easy integration of familiar graphical adjustments.
}

Example usage:
\dontrun{
plot(netlet, add_points = TRUE, point_size = 3) # Applies a fixed size to all nodes
plot(netlet, add_points = TRUE, point_size_var = "degree") # Sizes nodes based on their 'degree' attribute
}
}
