% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actorStats.R
\name{plot_actorStats}
\alias{plot_actorStats}
\title{Plotting function for actor level statistics for netify objects}
\usage{
plot_actorStats(
  summary_df,
  net_stat,
  longitudinal = FALSE,
  actor_names = NULL,
  toplist_n = NULL
)
}
\arguments{
\item{summary_df}{dataframe produced by summary_actor}

\item{net_stat}{character: actor-level statistic to be plotted (e.g., 'degree')}

\item{longitudinal}{logical: TRUE if longitudinal network, default is FALSE}

\item{actor_names}{character: actor name(s) in the network, default is NULL}

\item{toplist_n}{numeric: the number of actors to be listed as top n actors,
default is 20}
}
\value{
a graph object summarizing actor level stats of the network(s)
}
\description{
\code{plot_actorStats} takes in the output from \code{summary_actor} function
and outputs a graph with actor level statistics.
}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
