% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_stats.R
\name{plot_graph_stats}
\alias{plot_graph_stats}
\title{Plotting function for graph level statistics for netify objects}
\usage{
plot_graph_stats(summary_df, type = "line", specific_stats = NULL)
}
\arguments{
\item{summary_df}{A data frame produced by the \code{summary.netify} function
or any other function that outputs network statistics in a similar format.
The data frame should have a column "net" which identifies the network
or time point for each row.}

\item{type}{A character string specifying the type of plot to generate:
either 'line' for line plots or 'bar' for bar plots. The default is 'line'.}

\item{specific_stats}{Optional vector of column names from \code{summary_df}
specifying which statistics to plot. If NULL (the default), all statistics
in \code{summary_df} will be plotted.}
}
\value{
A \code{ggplot} object displaying the specified statistics for each network.
The plots are faceted by statistic type across the networks or time points.
}
\description{
\code{plot_graph_stats} takes a data frame containing graph level statistics and
generates a visual representation using either line or bar plots. This function
is designed to work with data frames that represent network statistics over
multiple networks or time points.
}
\author{
Ha Eun Choi, Cassy Dorff, Shahryar Minhas
}
