% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{plot_with_style}
\alias{plot_with_style}
\title{Apply style to netify plot}
\usage{
plot_with_style(netlet, style_fun, ...)
}
\arguments{
\item{netlet}{A netify object}

\item{style_fun}{A style function (e.g., style_budapest)}

\item{...}{Additional plot parameters}
}
\value{
A ggplot object
}
\description{
Helper function to properly apply styles to netify plots
}
