% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_igraph.R
\name{prep_for_igraph}
\alias{prep_for_igraph}
\alias{convert.to.igraph}
\title{Convert conflictNet object to igraph object}
\usage{
prep_for_igraph(netlet, add_nodal_attribs = TRUE, add_dyad_attribs = TRUE)
}
\arguments{
\item{netlet}{An R object}

\item{add_nodal_attribs}{Add any nodal attributes from netlet to igraph object, default is TRUE.}

\item{add_dyad_attribs}{Add any dyad attributes from netlet to igraph object, default is TRUE.}
}
\value{
igraph object
}
\description{
Convert conflictNet object to igraph object
}
\examples{

# load data
data(icews)

# cross-sectional case
icews_10 <- icews[icews$year==2010,]

# create netify object
dvars = c( 'matlCoop', 'verbConf', 'matlConf' )
nvars = c( 'i_polity2','i_log_gdp', 'i_log_pop' )
verbCoop_net = netify( 
  icews_10,
  actor1='i', actor2='j', 
  symmetric=FALSE, 
  weight='verbCoop',
  dyad_vars = dvars,
  dyad_vars_symmetric=rep(FALSE, length(dvars)),
  nodal_vars = nvars )

# convert to igraph object
igrph <- prep_for_igraph(verbCoop_net)
igrph

# longitudinal case
verbCoop_longit_net = netify(
  icews,
  actor1='i', actor2='j', time='year',
  symmetric=FALSE, 
  weight='verbCoop',
  dyad_vars = dvars,
  dyad_vars_symmetric=rep(FALSE, length(dvars)),
  nodal_vars = nvars )

# convert to igraph object
igrph_longit <- prep_for_igraph(verbCoop_longit_net)

# output in the longitudinal case is 
# a list of igraph objects
class(igrph_longit)
names(igrph_longit)
igrph_longit[['2002']]

}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
\keyword{conflictNet}
