% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.netify.R
\name{print.netify}
\alias{print.netify}
\title{Print method for netify objects}
\usage{
\method{print}{netify}(x, ...)
}
\arguments{
\item{x}{object of class netify, produced by get_adjacency}

\item{...}{additional parameters not used}
}
\value{
text information about the network
}
\description{
\code{print.netify} takes in a netify object
and defines information to be displayed on print.
}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
