% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{repeat_dyads_check}
\alias{repeat_dyads_check}
\title{Check whether dyadic observations are repeating in the data.frame object}
\usage{
repeat_dyads_check(dyad_data, actor1, actor2, time = NULL)
}
\arguments{
\item{dyad_data}{A data.frame containing the dyadic data.}

\item{actor1}{Character string specifying the column name for actor1.}

\item{actor2}{Character string specifying the column name for actor2.}

\item{time}{Optional character string specifying the column name for time.
If not provided, dyads are considered without regard to time.}
}
\value{
An integer count of the number of repeating dyads in the data.frame.
}
\description{
This function checks for repeating dyadic observations in a data.frame,
possibly considering a time dimension. It leverages fast C++ code for improved
performance.
}
\author{
Shahryar Minhas
}
