% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_bipartite_network}
\alias{style_bipartite_network}
\title{Preset style for bipartite networks}
\usage{
style_bipartite_network()
}
\value{
List of plot arguments
}
\description{
A complete visual style optimized for displaying bipartite networks
with two distinct node sets
}
