% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_cyberpunk}
\alias{style_cyberpunk}
\title{Cyberpunk-inspired network style}
\usage{
style_cyberpunk()
}
\value{
List of plot arguments
}
\description{
A futuristic network style with neon colors on dark background
}
