% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_dark2}
\alias{style_dark2}
\title{ColorBrewer Dark2-based network style}
\usage{
style_dark2()
}
\value{
List of plot arguments
}
\description{
A network style using ColorBrewer's Dark2 palette
for high contrast and accessibility
}
