% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_solarized}
\alias{style_solarized}
\title{Solarized-inspired network style}
\usage{
style_solarized()
}
\value{
List of plot arguments
}
\description{
A network style based on the popular Solarized color scheme,
optimized for reduced eye strain
}
