% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_temporal_network}
\alias{style_temporal_network}
\title{Preset style for temporal/longitudinal networks}
\usage{
style_temporal_network()
}
\value{
List of plot arguments
}
\description{
A complete visual style optimized for displaying networks over time
}
