% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_styles.R
\name{style_tufte}
\alias{style_tufte}
\title{Minimal Tufte-inspired network style}
\usage{
style_tufte()
}
\value{
List of plot arguments
}
\description{
A minimalist network style inspired by Edward Tufte's
data visualization principles
}
