% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.netify.R
\name{summary.netify}
\alias{summary.netify}
\title{Summary method to get graph level statistics for netify objects}
\usage{
\method{summary}{netify}(object, ...)
}
\arguments{
\item{object}{object of class netify, produced by get_adjacency}

\item{...}{additional parameters not used}
}
\value{
a data.frame object summarizing graph level stats of the network(s)
}
\description{
\code{summary.netify} takes in a netify object
and outputs a data.frame with graph level statistics.
}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
