% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_actor.R
\name{summary_actor}
\alias{summary_actor}
\title{Summary method to get actor level statistics for netify objects}
\usage{
summary_actor(netlet, ...)
}
\arguments{
\item{netlet}{object of class netify, produced by get_adjacency}

\item{...}{additional parameters not used}
}
\value{
a data.frame object summarizing actor level stats of the network(s)
}
\description{
\code{summary_actor} takes in a netify object
and outputs a data.frame with actor level statistics.
}
\author{
Ha Eun Choi, Shahryar Minhas
}
