% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_netify.R
\name{theme_netify}
\alias{theme_netify}
\title{theme_netify function}
\usage{
theme_netify()
}
\value{
A customized theme object for netify plots.
}
\description{
This function returns a customized theme for netify plots.
It is based on the \code{theme_minimal} function from the \code{ggplot2} package.
It removes axis text and titles from the plot.
}
\author{
Cassy Dorff, Shahryar Minhas
}
