% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_stat_netify}
\alias{theme_stat_netify}
\title{theme_stat_netify function}
\usage{
theme_stat_netify()
}
\value{
A customized theme object for netify stat plots.
}
\description{
This function returns a customized theme for netify stat plots.
}
\author{
Cassy Dorff, Shahryar Minhas
}
