% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_raw.R
\name{unnetify}
\alias{unnetify}
\title{unnetify}
\usage{
unnetify(netlet)
}
\arguments{
\item{netlet}{a netify object}
}
\value{
a list, matrix, or other object with netify attributes stripped
}
\description{
This function serves as a wrapper for
get_raw and other conversion functions
that convert a netify object back into
constituent parts
}
\author{
Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas
}
