% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnetify.R
\name{unnetify}
\alias{unnetify}
\title{unnetify}
\usage{
unnetify(netlet, remove_zeros = FALSE)
}
\arguments{
\item{netlet}{a netify object}

\item{remove_zeros}{Logical. If TRUE, remove edges with zero values.}
}
\value{
a data.frame object
}
\description{
This function converts a netify object and
any associated attributes into a dyadic level
data.frame
}
\author{
Cassy Dorff, Shahryar Minhas
}
