% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netify.R
\name{validate_netify}
\alias{validate_netify}
\title{Validates a new netlet object}
\usage{
validate_netify(netlet)
}
\arguments{
\item{netlet}{netlet object}
}
\value{
netlet
}
\description{
\code{validate_netify} is a helper function that validates inputs for a user-created netlet.
Not visible to users.
}
